+++ TR5Main/Objects/TR2/Entity/tr2_barracuda.cpp
@@ -32,8 +32,8 @@ void BarracudaControl(short itemNumber)
 	auto* item = &g_Level.Items[itemNumber];
 	auto* creature = GetCreatureInfo(item);
 
-	short angle = 0;
-	short head = 0;
+	float angle = 0;
+	float head = 0;
 
 	if (item->HitPoints <= 0)
 	{
