+++ TR5Main/Objects/TR2/Entity/tr2_mercenary.cpp
@@ -34,12 +34,12 @@ void MercenaryUziControl(short itemNumber)
 	auto* item = &g_Level.Items[itemNumber];
 	auto* creature = GetCreatureInfo(item);
 
-	short angle = 0;
-	short tilt = 0;
-	short headX = 0;
-	short headY = 0;
-	short torsoX = 0;
-	short torsoY = 0;
+	float angle = 0;
+	float tilt = 0;
+	float headX = 0;
+	float headY = 0;
+	float torsoX = 0;
+	float torsoY = 0;
 
 	if (item->HitPoints <= 0)
 	{
@@ -205,12 +205,12 @@ void MercenaryAutoPistolControl(short itemNumber)
 	auto* item = &g_Level.Items[itemNumber];
 	auto* creature = GetCreatureInfo(item);
 
-	short angle = 0;
-	short tilt = 0;
-	short headX = 0;
-	short headY = 0;
-	short torsoX = 0;
-	short torsoY = 0;
+	float angle = 0;
+	float tilt = 0;
+	float headX = 0;
+	float headY = 0;
+	float torsoX = 0;
+	float torsoY = 0;
 
 	if (item->HitPoints <= 0)
 	{
