+++ TR5Main/Objects/TR2/Entity/tr2_worker_flamethrower.cpp
@@ -29,7 +29,7 @@ enum WorkerFlamethrowerAnim
 
 };
 
-static void Flame(int x, int y, int z, int velocity, short yRot, short roomNumber)
+static void Flame(int x, int y, int z, int velocity, float yRot, short roomNumber)
 {
 	short fxNumber = CreateNewEffect(roomNumber);
 	if (fxNumber != NO_ITEM)
@@ -70,12 +70,12 @@ void WorkerFlamethrower(short itemNumber)
 	auto* item = &g_Level.Items[itemNumber];
 	auto* creature = GetCreatureInfo(item);
 
-	short tilt = 0;
-	short angle = 0;
-	short headX = 0;
-	short headY = 0;
-	short torsoX = 0;
-	short torsoY = 0;
+	float tilt = 0;
+	float angle = 0;
+	float headX = 0;
+	float headY = 0;
+	float torsoX = 0;
+	float torsoY = 0;
 
 	pos.x = WorkerFlamethrowerBite.x;
 	pos.y = WorkerFlamethrowerBite.y;
