+++ TR5Main/Objects/TR2/Entity/tr2_yeti.cpp
@@ -47,10 +47,10 @@ void YetiControl(short itemNumber)
 
 	bool laraAlive = LaraItem->HitPoints > 0;
 
-	short angle = 0;
-	short torso = 0;
-	short head = 0;
-	short tilt = 0;
+	float angle = 0;
+	float torso = 0;
+	float head = 0;
+	float tilt = 0;
 
 	if (item->HitPoints <= 0)
 	{
