+++ TR5Main/Objects/TR4/Entity/tr4_big_beetle.cpp
@@ -47,7 +47,7 @@ namespace TEN::Entities::TR4
 		auto* item = &g_Level.Items[itemNumber];
 		auto* creature = GetCreatureInfo(item);
 
-		short angle = 0;
+		float angle = 0;
 
 		if (item->HitPoints <= 0)
 		{
