+++ TR5Main/Objects/TR4/Entity/tr4_troops.cpp
@@ -68,12 +68,12 @@ void TroopsControl(short itemNumber)
 	auto* creature = GetCreatureInfo(item);
 	auto* object = &Objects[item->ObjectNumber];
 	
-	short angle = 0;
-	short tilt = 0;
-	short rot = 0;
-	short joint0 = 0;
-	short joint1 = 0;
-	short joint2 = 0;
+	float angle = 0;
+	float tilt = 0;
+	float rot = 0;
+	float joint0 = 0;
+	float joint1 = 0;
+	float joint2 = 0;
 	
 	int dx = 0;
 	int dy = 0;
