+++ TR5Main/Objects/TR4/Entity/tr4_wildboar.cpp
@@ -33,14 +33,14 @@ void WildBoarControl(short itemNumber)
 	auto* item = &g_Level.Items[itemNumber];
 	auto* creature = GetCreatureInfo(item);
 
-	short angle = 0;
-	short head = 0;
-	short neck = 0;
-	short tilt = 0;
-	short joint0 = 0;
-	short joint1 = 0;
-	short joint2 = 0;
-	short joint3 = 0;
+	float angle = 0;
+	float head = 0;
+	float neck = 0;
+	float tilt = 0;
+	float joint0 = 0;
+	float joint1 = 0;
+	float joint2 = 0;
+	float joint3 = 0;
 
 	if (item->HitPoints > 0)
 	{
